feat 이미지 첨부 리뷰 작성 기능 완성
UUID로 이름 변경

+++ src/main/java/com/multi/culture_link/festival/model/dto/VWUserReviewDataDTO.java
@@ -8,7 +8,7 @@ public class VWUserReviewDataDTO {
 	private int festivalReviewId;
 	private int userId;
 	private int festivalId;
-	private int festivalReviewStar;
+	private double festivalReviewStar;
 	private String festivalReviewContent;
 	private String attachment;
 	private String userName;
